Move IsNull/IsNotNull predicates to QL project
